Change LUT size in Benchmark from 16-Bit to 32-Bit
This changes the size of the LUT used in the Benchmark suite from 16 to
the new 32-Bit size.

fixes #18

+++ benchmarks/conftest.py
@@ -29,7 +29,7 @@ def lut():
     x = np.arange(256)
     x[x >= 128] -= 256
     xx, yy = np.meshgrid(x, x)
-    return torch.from_numpy(xx * yy).short()
+    return torch.from_numpy(xx * yy).int()
 
 
 @pytest.fixture(params=networks)
